Refactor all files in project
