--- .funcignore
@@ -1,9 +1,0 @@
-.git*
-.vscode
-__azurite_db*__.json
-__blobstorage__
-__queuestorage__
-local.settings.json
-test
-README.md
-.venv
