--- MainDirectory/database/drop_tables.sql
@@ -1,20 +1,0 @@
--- DROP all tables in database
-
-DROP TABLE instruction_steps;
-DROP TABLE steps;
-
-DROP TABLE user_progresses;
-
-DROP TABLE reviews;
-
-DROP TABLE recipe_ingridients;
-DROP TABLE ingridients;
-
-DROP TABLE recipes;
-
-DROP TABLE recipe_details;
-
-DROP TABLE users;
-
-DROP TABLE categories;
-
