+++ app/database/drop_tables.sql
@@ -1,0 +1,19 @@
+-- DROP all tables in database
+
+DROP TABLE [dbo].[instruction_steps];
+DROP TABLE [dbo].[steps];
+
+DROP TABLE [dbo].[user_progresses];
+
+DROP TABLE [dbo].[reviews];
+
+DROP TABLE [dbo].[recipe_ingridients];
+DROP TABLE [dbo].[ingridients];
+
+DROP TABLE [dbo].[recipes];
+
+DROP TABLE [dbo].[recipe_details];
+
+DROP TABLE [dbo].[users];
+
+DROP TABLE [dbo].[categories];
