+++ app/services/user_service.py
@@ -1,0 +1,36 @@
+import jwt
+from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
+from datetime import datetime
+from pydantic import BaseModel
+from datetime import datetime, timedelta
+
+SECRET_KEY = "secret_key" #TODO: hide !!
+
+ACCESS_TOKEN_EXPIRE_MINUTES = 30#minutes 
+security = HTTPBearer()
+
+class UserService:
+    
+    @staticmethod
+    def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
+        try:
+            token = credentials.credentials
+            payload = jwt.decode(token, SECRET_KEY, algorithms=["HS256"])
+            username = payload.get("sub")
+            
+            if username is None:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
+            
+            return token
+        except jwt.ExpiredSignatureError:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token has expired")
+        except jwt.PyJWTError:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
+
+    @staticmethod
+    def create_jwt_token(data: dict, expires_delta: timedelta = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)) -> str: #returns token
+        to_encode = data.copy()
+        expire = datetime.utcnow() + expires_delta
+        to_encode.update({"exp": expire})
+        return jwt.encode(to_encode, SECRET_KEY, algorithm="HS256")
