--- function_app.py
@@ -1,5 +1,0 @@
-import azure.functions as func
-
-from main import app as fastapi_app
-
-app = func.AsgiFunctionApp(app=fastapi_app, http_auth_level=func.AuthLevel.ANONYMOUS)
