--- host.json
@@ -1,20 +1,0 @@
-{
-  "version": "2.0",
-  "extensions": {
-    "http": {
-        "routePrefix": ""
-    }
-  },
-  "logging": {
-    "applicationInsights": {
-      "samplingSettings": {
-        "isEnabled": true,
-        "excludedTypes": "Request"
-      }
-    }
-  },
-  "extensionBundle": {
-    "id": "Microsoft.Azure.Functions.ExtensionBundle",
-    "version": "[4.*, 5.0.0)"
-  }
-}
