--- main.py
@@ -1,36 +1,0 @@
-import azure.functions as func
-import fastapi
-
-from MainDirectory.routers.recipe_router import recipe_router
-from MainDirectory.routers.user_router import user_router
-from MainDirectory.routers.recipe_ingredients_router import recipe_ingredient_router
-
-from MainDirectory.database.database import create_database
-
-create_database()
-
-app = fastapi.FastAPI()
-
-app.include_router(recipe_router)
-app.include_router(user_router)
-app.include_router(recipe_ingredient_router)
-
-@app.get("/")
-async def hello():
-    return {
-        "Hello" : "world !"
-    }
-
-
-@app.get("/sample")
-async def index():
-    return {
-        "info": "Try /hello/Shivani for parameterized route123.",
-    }
-
-
-@app.get("/hello/{name}")
-async def get_name(name: str):
-    return {
-        "name": name,
-    }
